+++ tests/derive-tests/as_variant/pass/enum/zero_variants.rs
@@ -1,6 +1,0 @@
-use enumcapsulate::derive::AsVariant;
-
-#[derive(AsVariant)]
-pub enum Enum {}
-
-fn main() {}