+++ tests/derive-tests/as_variant_mut/fail/enum/tuple_variants/multiple_fields.out.rs
@@ -1,6 +1,0 @@
-use enumcapsulate::derive::AsVariantMut;
-pub struct VariantA;
-pub enum Enum {
-    VariantA(VariantA),
-    VariantB(i32, u32),
-}