+++ tests/derive-tests/as_variant_mut/fail/enum/unit_variants.rs
@@ -1,9 +1,0 @@
-use enumcapsulate::derive::AsVariantMut;
-
-pub struct VariantA;
-
-#[derive(AsVariantMut)]
-pub enum Enum {
-    VariantA(VariantA),
-    VariantB,
-}