+++ tests/derive-tests/as_variant_mut/pass/enum/mixed_variants.rs
@@ -1,0 +1,28 @@
+use enumcapsulate::derive::AsVariantMut;
+
+pub struct VariantA;
+pub struct VariantB;
+
+#[derive(AsVariantMut)]
+pub enum Enum {
+    Unit,
+    ZeroTupleFields(),
+    ZeroStructFields {},
+    OneTupleField(VariantA),
+    OneStructField { variant: VariantB },
+    TwoTupleFields(i32, u32),
+    TwoStructFields { a: i32, b: u32 },
+}
+
+fn main() {
+    use enumcapsulate::AsVariantMut;
+
+    let mut subject = Enum::Unit;
+
+    {
+        let _: Option<&mut VariantA> = subject.as_variant_mut();
+    }
+    {
+        let _: Option<&mut VariantB> = subject.as_variant_mut();
+    }
+}