+++ tests/derive-tests/as_variant_ref/fail/enum/tuple_variants/zero_fields.out.rs
@@ -1,6 +1,0 @@
-use enumcapsulate::derive::AsVariantRef;
-pub struct VariantA;
-pub enum Enum {
-    VariantA(VariantA),
-    VariantB(),
-}