+++ tests/derive-tests/from_variant/fail/enum/unit_variants.rs
@@ -1,9 +1,0 @@
-use enumcapsulate::derive::FromVariant;
-
-pub struct VariantA;
-
-#[derive(FromVariant)]
-pub enum Enum {
-    VariantA(VariantA),
-    VariantB,
-}