+++ tests/derive-tests/variant_downcast/fail/enum/tuple_variants/zero_fields.rs
@@ -1,9 +1,0 @@
-use enumcapsulate::derive::AsVariant;
-
-pub struct VariantA;
-
-#[derive(AsVariant)]
-pub enum Enum {
-    VariantA(VariantA),
-    VariantB(),
-}